Remove problematic count() in MediaInfoPrefetchingTermLookup
There is no need to actually *count* how many values these arrays
contain, if all we need to know is *if* the array is (not) empty.

The first variable is guaranteed to be an array. In this case, we can
use PHP's magic to-bool cast that's automatically happening. If you
prefer, we can also write `$diffTermTypes !== []`.

The second variable might be null. We don't want the if() to succeed
when it's null (this is what the is_array check did before). Therefor
we do a strict comparison with, and only with the empty array.

Bug: T239165
Change-Id: I157aae6ac922695262aa7be6604cc96fe150f7c5