Replace oldskool @see PHPDoc tags with @inheritDoc
We started using @see tags like this in the Wikibase codebase in 2014
to point to the parent documentation, as well as to mark such methods as
being documented. (Just not having a PHPDoc block would look like it was
forgotten.) Much later, I think in 2018, we started using @inheritDoc in
all codebases to do the same.

Change-Id: I45c8730a5a80e91c01d33563a3b54a1d94c7d6e0