Address ESLint max-len warnings
Currently ESLint spits out a ton of max-len violation warnings; these reduce the
signal-to-noise ratio in code editors and command-line test output.

I've tweaked the linter rules in the following ways to better reflect how we're
actually writing code now:

* Max length for code has been bumped up to 110 (looking at you, ES5...)
* Added the "ignore comments" option
* Disabled max-len in test files entirely, since QUnit can get pretty verbose

I've also modified the code in a few places where it seemed like changes could
be introduced to improve legibility.

There are still a few files (currently 4) that violate the updated rules. But
this is still an improvement in my opinion.

Change-Id: Ic2b5689e75da1fc32d75b6e2044ab1a5e63576e1