Escape labels in HTML output
I suspect this slipped in because I at some point copied markup
from another template, where {{{label}}} is unescaped because it's
the html-formatted version (e.g. linking out to the item page)

In the search autocomplete, however, the label should always be
escaped. That label is always simply the plaintext item label, and
should not, ever, be parsed.

Bug: T240773
Change-Id: If47903f11645e52d56a875be4c03de47400cef01