Show wbparsevalue errors in the WBMI UI
When a user adds a statement value, an API call to wbparsevalue
may return an error message, e.g. for a URL with a missing or
invalid scheme. This commit adds this error message to the UI
as an OOUI MessageWidget.

This also adds a method to some input widgets to flag their input
as invalid using the OOUI setValidityFlag method.

Finally, this abstracts out the setError method to the
ComponentWidget class so we can use it in the StatementInputWidget
here and, in a future patch, in the ItemWidget and StatementWidget.

Bug: T239857
Change-Id: I55e02861363a2f1b0979e3d67f627addf83557c6