Minor JSDoc (-related) cleanup
I started out wanting to make config documentation similar across
files. `@param config.` is currently more prevalent, so I decided
to make `@cfg` docs the same.

While going over those, I discovered a couple of arguments that
were incorrectly documented as optional ([]) when they aren't, or
vice versa.

Also found an argument that is no longer in use, and one that we
forgot to pass on to the child object that actually uses it.

Change-Id: I23aa66102c687573e39157c6306b550f87bdaf8d