Update error UX on submit
This patch catches errors in the StatementWidget.submit() method,
saves them in state, and adds them as OOUI MessageWidgets to the
template. This patch makes it possible to show multiple errors at
once for the StatementWidget and other widgets that currently show
errors (StatementInputWidget) or will, in the future, show errors
(ItemWidget).

Because we can't reliably differentiate item-specific and non-item-
specific errors based on the API response for wbsetclaim, for now
we will show all errors at the top of the StatementWidget.

Bug: T233051
Change-Id: I7c50e3aa677d224180f67b03720771c9733679f3