Use valueType instead of propertyType
Since Ib7fedbd1447140cc7f118916af3805e02bbd902e, wbparsevalue will
also accept the property id instead of the property datatype.
This allows us to get rid of the property datatype: this was the only
place we really needed it for, and we already have the property ID
anyway - it's an easy replacement.

We now no longer need propertyId for any of a statement's internals,
so let's switch to accepting the valueType. Reason:
propertyType might be hard to come by, as it's not part of any of
the datamodel objects or the serialized JSON. The value type & the
property ID OTOH are! So if we can get by with only these 2, it lowers
the barriers of creating a StatementWidget instance.

This is useful for e.g. UploadWizard's 'copy-statement' functionality
(currently broken) that reads and existing statement's datavalue object
and constructs another one based off of that. It makes things much
simpler not having to find the property datatype, but just being able
to use the data already available...

Bug: T232499
Change-Id: I0de6339769275cda4baca168f41ba1eb32c11a79