Add a (default) collapsed version for globecoordinate input
The map is massively useful to input coordinates, but also very
distracting from other data on this page.

Most page views will not be about adding data, but consuming it,
and we don't want to distract users from that data when they have
no intention of adding coordinates.

This patch adds a 1-line version of the coordinate input. It's
nowhere near as fancy as the full-blown map input, but it does
have a button to expand to the full experience.

Distraction will be minimal for casual viewers (just a 1-line
input area like other input types), but those that do intend to
input the data can still go all the way.

- Also included a rendering fix in ComponentWidget - will also
submit that in a separate patch & rebase this one later on.

Bug: T227116
Change-Id: I3f3394cf6661edb7a056583477a073de0466c66e