Component rerender reconciliation performance refactor
The component rerender process is a massively complex routine
where an attempt is made to find matching nodes between the
currently displayed DOM and the newly generated DOM, in order
to merge them together in a way that no state is lost from
existing nodes (e.g. focus or input)

This basically means that it'll grow an order of magnitude
for each additional node, because every node needs to be
checked against every other node (roughly speaking)
We've not had any performance issues with this so far, but
that's only because our widgets are quite small. Given a
few thousand nodes to rerender, our current routine would
start to choke.

This is a minor performance refactor. There are some
optimizations on how to find matching nodes (e.g. don't
compare children if the parents are a different tag),
and how to merge the DOM (e.g. get rid of terminato node)
but the biggest difference, by far, in this refactor is that
we'll no longer try to match up all nodes - only the ones
that have a state that would get lost upon replacing them:
- OOUI and other nodes used elsewhere in the code
- inputs with manual input
- the active/focused node
This is a finite list under user control, so this routine
now scales linearly.

This also fixes a theoretical bug where DOM might not fully
rebuild properly when there are drastic differences between
the old & new DOM: old code was iterating the array of
matched nodes instead of the new DOM, so if the size of them
is different enough, and position of non-existing nodes in
the new DOM exceeds the old length, some old nodes might
have remained. (I only accidentally discovered this now that
I'm drastically reducing the amount of matched nodes but,
while it has not yet occured with our existing templates, it
is could possibly happen already)

Change-Id: I558803fe971ef1a3f956fbce81f2e9cc468dec28