Separate autocomplete element from entity input widget
Sadly, after recent efforts to consolidate 3 different
entity input variation, I must already start to break
this apart.

The autocomplete input field is being separated form
the overarching EntityInputWidget because this specific
little thing will have to be implemented elsewhere, as
part of the quantity input widget (to select entities
to serve as unit)

I've tried to separate things to line up with other
current practices: the autocomplete thing, as far as
I'm concerned is similar to an OO.ui.TextInputWidget
or so, whereas the EntityInputWidget is the more
advanced thing which utilises the input field.

Meanwhile I also made a minor adjustment to
AddPropertyWidget, which had become misaligned due to
the change in DOM element order (the autocomplete
is now nested deeper inside the EntityInputWidget,
instead of *being* the widget)

And with EntityInputWidget no longer being a
TextInputWidget, it no longer had a focus method. I
figured that'd be a decent addition to the base widget
interface as well.

Change-Id: Iec096e2fc2efaf1ab6aa065763ef29a0e0799a19