wikibase/data-model-services": "~4.0|~5.0"
This makes the tests calling the changed public interface
compatible with the 5.0 release.
There is one other call to the breaking change code but it
doesn't need to be updated.

This leaves 1 implementation of the interface with defaults, that will
throw and exception if used, so we can stay compat with both 4.0 and 5.0
for a short period of time during migration.

Bug: T243596
Change-Id: Ice9b4c03274a8a375bc97b33eeced971435a3bd2