Implement support for time datavalue
This includes some changes in AbstractInputWidget.parseValue,
which was previously only used for validation.
Now that it's also used to actually parse input, we need:
- to be able to abort in-flight requests when input has changed
- be able to parse when property is not known
- stop the client-side only validation call (we can't validate
  data on the client yet, because we need the parseValue call
  to build the data in the first place)

And we'll now (again) have to also prepopulate the formatValue
cache with plaintext versions for the qualifiers: the plaintext
version of an existing time qualifier will be used in the time
input field.

Bug: T231979
Change-Id: I1364a2f19e7630224196a54f284458b3d90852ba