Default to unsupported input unless instructed otherwise
These inputs are currently being constructed in 2 different ways:
- new statement-level inputs are constructed with a specific type,
  so this change has no effect
- qualifiers are constructed without type and receive their type
  later on via `setData` (if there is existing data) or via
  `setInputType` (when the property changes)

In this latter case, it doesn't make sense to initialize with an
empty string - it'll change anyway. There is a downside to starting
with an empty string, though: QualifierWidget.asyncFormatForDisplay
will attempt to format that empty string when state changes (which
it does when toggling edit mode). Since there is no actual data
there, it makes no sense to pretend that we have an empty string
and waste API calls trying to format it, when in reality, at that
point, we just don't have valid data...

Change-Id: I45ad93822640f93e67c526da38287d18b74b5243