Add somevalue/novalue options for entity datatype
This patch:
- Adds a dropdown that allows the user to select the snak type
and supports adding somevalue and novalue snaks via the UI.
Currently this is only enabled for properties with the Wikibase
entity datatype, but it could be extended to all datatypes. This
would likely require some design changes.
- Displays existing somevalue and novalue snaks for all datatypes,
which could in theory be added via the API. For somevalue snaks
we are only supporting the use case of missing Wikidata items, but
at least this way somevalue/novalue snaks for other datatypes added
via the API won't be hidden from users, and we will know if users
end up using other snak types for other datatypes

Bug: T239172
Change-Id: I6afac6e0d302bf794c971f63f7dd19bd12ca2933