Ensure correct entity id is used when choosing between same-label options
A entity id => label map is used to allow textual input to match
known entities (e.g. selecting something from autocomplete, deleting
a character from input field, and then adding it again to form the
same input that was selected)

However, this defaults to the first entity id that matches the
label. That's usually fine, except when there are multiple same
labels: even though we know the entity id, it'll default to using
the code path that selects the first matching entry.

This is easily worked around by explicitly overriding the entity
id *after* updating the input field text to the label.

Bug: T246918
Change-Id: Ibd8b34789dfeadd0c2673b39050aa277d1a946d4