Pick up on same-claim-different-guid differences
Scenario:
- have a an existing set of statements: e.g. depicts=cat, depicts=dog
- add a new one that already exists, e.g. another depicts=cat
- cancel (which should return you back to the original depicts=cat, depicts=dog)
- find that the new depicts=cat is also present in read mode

Above scenario currently happens because when we're checking for removals,
GUIDs are not being compared, only the actual "content". Since both
depicts=cat entries have the same content (same claim, references, rank, ...),
we don't notice that the last one was actually a new addition, and should
be discarded on cancel.

Bug: T247081
Change-Id: I75eddd58e6e0585f13abf1c342c1a7277c460eae