Prevent coordinates from being re-arranged in RTL mode
Coordinate strings will contain both numeric and non-numeric symbols. We need
to override the standard RTL way of displaying this content to prevent the
strings from being re-arranged into nonsense values. This is achieved by adding
a direction: ltr @noflip rule to the coordinate input, and wrapping the
ItemWidget label in a <bdi> tag.

On the PHP side, the createFormattedDataValue method has also been updated to
wrap the resulting value markup in a bdi tag.

Bug: T246797
Change-Id: Ie772eef6e091131128fd262b0a8fca49cafacae0