Get rid of &#8232; character
The actual separator between property & value is `: ` (colon + space)

Setting that up (in a somewhat readable way) in HTML/mustache template
means having some whitespace before and after the separator node.
And those, of course, show up (collapsed into a single space) in the UI.

In order to get rid of that unwanted whitespace, we'd apply `display: flex`
to the container element, but that also meant that the intended space
in the separator also got collapsed. To tackle that, we added &#8232;
which should be a non-printing character: that'd then protect the
intentional space from being collapsed.

We've just recently changed that `display: flex` to `display: inline-table`
(in If18b60858f09ac69e77912ba461cd550aefd1dc3, for other reasons),
which interprets whitespace differently - the way we actually want it to
be displayed - and we no longer need the non-printing character hack.

Bug: T245681
Change-Id: I5606ec01228cd066c34d09d181b231bf76e8e26a