Improve handling of invalid qualifiers in GlobeCoordinateInputWidget
Changes the way that data is set and "change" events are emitted in the event
of invalid data (i.e. when parseValue call fails). Previously, coordinate data
would be reset to null if parseValue failed and changes would be emitted
regardless of success or failure. This could lead to cryptic behavior from the
user's perspective; if the user changed a previously-valid coordinate qualifier
value to something invalid and then clicked "publish", the (now-null) qualifier
would silently disappear. This seems like it could be pretty confusing.

This change removes the call to setState() in the parseValue catch handler
entirely, so any previously-valid data is kept if new data is invalid.
Likewise, change events are only emitted when parseValue is successful.

If the user's only change is to modify a previously-valid qualifier to a set of
invalid coordinates, they will now be unable to publish such a change.

If they also make other changes in the same widget (adding new items, or
adding/changing other qualifiers), then the previous qualifier value will be
kept, not discarded.

Clearing the coordinates from a qualifier will still behave normally.

Bug: T233052
Change-Id: Id2d26d0965e4e4fd16b9548728c2cd6c77474f19