Prefer detailed error message over plain error response
I made a change recently to prefer the plain `error.info` response
over the more complex HTML version provided by RepoApiError.
Looks like that was a bad call, because the latter may print
more elaborate error messages, when there is more than one.

It looks like parseValue, which was the reason for using the plain
message in the first place, can even use RepoApiError as well
(IDK if this was an oversight when this was introduced, or if
things have changed since, but error responses seem the same)

And we don't need the 'abort' special case anymore AFAICT:
- the inputs using parseValue don't check the error response
- the parseValue-based error reporting (RepoApiError) will
  fall back to reading error.exception, which will include
  the term 'abort' (and 'abort' is not something a user
  should see anyway; it should be handled in code, if relevant,
  and is easy to check from the errorCode, which is also 'abort')

So... this restores things to the more complete RepoApiError
responses, and changes parseValue error reporting to be consistent
with this (instead of the other way around)

Bug: T233051
Change-Id: I2004c9ead399f27589189d784e864b07388532ff