Improved pending state for StatementPanel
Uses OO.ui.mixin.PendingElement to set a pending state when the
StatementPanel's sendData method is called. This state is removed in the
.always() handler of the returned promise.

Additionally adds a new CSS rule to hide elements in the qualifier area
(.wbmi-content-items and .wbmi-statement-footer) while the panel is in pending
state, so that users don't see the intermediate states of any QualifierWidgets
as data is being published.

Bug: T228947

Change-Id: I7af2bf4281f75eb5fc01e8f9bde4729d95e18eb7