Fix IE11 incompatibility
These 2 lines are basically checking whether a value is an
object literal (= `{}`). This check requires an actual value,
though: undefined (in case the key doesn't exist) will cause
a TypeError. So it'd fall back to an empty string in case of
undefined, which was guaranteed to make the comparison fail
(`Object.getPrototypeOf` of a string would return the String
prototype, not the Object prototype...)
And of course, the value could already be a string in the
first place.

IE11, however, appears to fail with strings.

I'll add an additional check for instanceof Object.
The `Object.getPrototypeOf` condition needs to stay because
we only care about object literal, not object instances of
whatever class.
This should work well enough for working around any (string
or other non-object) incompatibilities.

Bug: T249556
Change-Id: I0fc3b0967b4649adf672036ad62e6359d18ec835