Allow deleted properties
Right now, pages that have a deleted property will throw
an exception. This will accept deleted properties, and update
the frontend to handle the fact that we now no longer know
their datatype (essentially remaining functional so these
statements can be deleted)

Meanwhile also fixed formatValue handling: there was a built-in
assumption that formatting of entities would always return the
same, regardless of the property. That's no longer a valid
assumption, it seems: values associated with a nonexisting
property get an error message when formatted. I've updated
the merge handling and cache shortcut to address that reality.

Bug: T249688
Change-Id: Ie23d62f2ebb9b3f219afa1a7cbd319887a0c72b1