Optimize media search
Search requests are very slow at the moment, and the extended
metadata seems to be the culprit. Extended metadata is derived
from the page's content, and that appears to happen ad hoc,
a very intensive operation when it needs to be done for 40 files
at once...

So... I'm dropping the extended metadata for now: they're not
usable in their current implementation. This means we lose:
- enhanced filenames: not a big deal, most weren't that different
  from their filename anyway
- descriptions: not that bug of a deal, we already prefer
  captions (in local language) over (mostly english) wikitext
  excerpts anyway
- license information: not a dealbreaker, but I'm slightly sad
  that we're not able to show this immediately

And latest mockups didn't have any of the above anyway!

Bug: T216625
Change-Id: I44ca5ffedfd6406e58b6488a3879f9f90d605b2b