Image size filter
This is a JS-only feature (for now at least) because OOUI
does not natively support anything like this, AFAICT.
Once research has completed and we know what direction we
want to take this in, we can reconsider how all of this
needs to work.

Meanwhile, I've also changed the API a little: instead of
accepting the desired filetype, it'll now simply accept
a raw search string: this allows more flexibility as we
can now filter by filetype and/or resolution (ot others)

And I've made changes to infinite load/scroll. That was
a little flaky (and likely still is to some extent), but
it became increasingly troubling with another filter thrown
in the mix...

Bug: T250732
Change-Id: I4445a6da159b76a8dd8e77f3c0665eecfb5771eb