Add dummy concept chips without thumbnail
I suspect that fetching thumbnails for concept chips might
require significant additional resources: we'll likely
need to execute an actual search for each chip to find a
relevant thumbnail.

I'm not sure these tiny thumbnails provide enough added
benefit to warrant that additional load. It appears that
even Google isn't even sure about this (I don't get thumbs
in Google images, but have seen variations with thumbs for
other people)

Let's have a few dummies with and a few without, so that
we can research this and make a more informed decision.

Bug: T249614
Change-Id: Ia023d5b9b34bde4a9b2b84b341cc68d22fa75cf1