Follow-up 1827c7a: Ensure inNamespace() is called only on Title object
This was caused by 1827c7a, where the Wikibase Client's TitleFactory was just
replaced with core's TitleFactory, but they were not equal.

Precisely, Wikibase Client's returns Title or throws StorageException, while
core's return Title or null. FilePageLookup.php makes a call to inNamespace()
inside the try-block, and relied on the behavior of exception in such block
(i.e., an exception will force the inNamespace() call to be skipped). But the
switch to core's TitleFactory made the call return only Title and null, and
subsequent call to the inNamespace() method continued unconditionally, both
on Title and on null.

Bug: T253792
Change-Id: Ic9e8a211f3505c7f0f151a716df56cd2fde787a7