Merge config module into statements module
Of the modules that previously used the config module, one is the
statements module, and the other directly depends on it. This means that
we can move the config into the statements module, and shrink the
startup module metadata (which is served with all pages, not just the
ones where these modules are relevant) a little bit.

Bug: T254151
Change-Id: Ic51828799e12a9357db43ee88012aa62ff04a54d