Refactor Less variables
This patch:

- Sets up a system for importing Wikimedia UI base variables.
wikimedia-ui-base is included as a dev dependency and a new
npm script can be used to copy the variables file to a local
lib directory. The file is then imported in our extension-specific
variables file and will therefore be available in any file that
imports mediainfo-variables.less.
- Updates mediainfo variables to follow the
extension-property-block-element--modifier pattern. Prefixing
our variables with wbmi- makes it easy to tell if a variable is
extension-specific or from the WMUI base.
- Updates Less files to use updated variable names, or to use
WMUI base variables if they make sense semantically
- Updates mixin syntax so we no longer have to ignore
stylelint rules

Change-Id: Icf9a77c05856b3e4092909f632be941ecd11d75f