Use type=text input field for numeric types rather than type=number
type=number feels like it'd be the right one here, but it
always returns input with `.` as decimal separator, even when
another (e.g. `,` in German) is used.

Wikibase's parsevalue tries to interpret the input based on
the user's language, but the normalized numeric input will
never contain their actual decimal separator (`,`) because
it implicitly converts to `.`. As a result, it's not currently
possible to enter fractional values in languages where the
decimal separator is not `.`

type=number added some additional browser built-in protection
against invalid input (by not allowing characters that don't
make sense for numbers), but it leads to even worse results
(implicit conversion of decimal values) in combination with
wbparsevalue. Wikidata also uses type=text, so we should
probably just stick with that - wbparsevalue validates the
input anyway.

Bug: T254615
Change-Id: Ibaa45fe4ad9910e9460a7c6cc188d678aed7a659