Use auto flex-basis instead of 0, fixing flexbox issues in IE
.flex() defaults to using `auto` when the value is omitted,
but because we needed to pass the order arg, we also had to
set the flex-basis arg. I assume I set this to 0 because it's
the browser default, but the actual default for this .flex()
is `auto`. It turns out that `0` width is rendered differently
between IE and other browsers, so defaulting back to `auto`.

Bug: T249556
Change-Id: I912e580e69ab69bd9b9dd9a31bbfb33123e6ad52