Use fulltext search to find wikidata entities & weigh them differently
The existing method only accepted perfect matches. This was
intenional because any entity match would show up before any
other type of search result, and there was no way to weigh
multiple results differently.

We're in a totally new situation now, and we should be open
to more wikidata entities. E.g. searching for "flag russia"
should find the "flag of russia" entity. "cats" should find
"housecat", etc.

Opening up to more results also means that we might have more
barely-relevant results. There is no clear relevance score
available, though, so I'm hacking something together based on
the order of the results and the matched text in the highlight
query. It's far from perfect, but it's something, and
it's deterministic in how humans can enricht the data to
better serve searches - it's not a black box.

Bug: T258051
Change-Id: Ide44de27a5a9350c360105850ef26529fb990cf8