Make MediaSearch query scoring more similar to Cirrus defaults
Before this patch, both 'all' and individual fields were used
to score the query, so a match in any of those fields would
be ranked higher than it should (because it'll also match in 'all')
Like Cirrus, 'all' is now only used in the top-level filter,
and the individual fields are now used to generate the score.

I've updated those too, to be more like Cirrus. These queries
are now the same (some fields were missing, and there are some
tweaks like dis_max) and so are the weights. I'm sure we'll end
up having to tweak this, but having a similar starting point is
going to be useful to determine the impact of other changes.

Bug: T252685
Change-Id: I9f3e8f3e8bbabecc37b1385ead059da6433daa68