Recreate autocomplete search input
This patch adds a new base component, AutocompleteSearchInput,
which is based on the in-progress WVUI Input component, removes
some functionality and flexibility not needed for the Media Search
use case, and adds support for fetching, displaying, and allowing
the user to interact with autocomplete results. The process for
fetching results here is the same as it was in the previous
iteration of Media Search and all interactive and accessibility
features provided by the old OOUI LookupElement and
SearchInputWidget are supported.

Bug: T256673
Bug: T257693
Change-Id: Id0e7fdf09ef662d454f1ef924d7b47bb1624fd82