Fix array unpacking as argument list
If an empty array is unpacked into argument list with spread operator,
no value will be passed, so it cannot be used as an argument of a
function that requires at least one argument except if there's a
guarantee that the array to unpack will never be empty.

If MediaInfoEntityTermsView::getCaptionsContent() returns [], then no
value will be passed to Tag::appendContent(), which will generate error
since the function requires at least one argument at all times.

Bug: T259745
Change-Id: Ia0508c81651157ed051810057419f54c033554b1