Update Icon base component
I've pulled this out of the autocomplete work so other branches
can be based off of this before that one is merged.

This updates the Icon base component to closely mirror the new
WVUI Icon component, which produces an SVG icon. Because we don't
yet have access to the icon data contained in the WVUI library,
I've added a file with the icon data we need for Media Search. If
we ever swap out our base components for the WVUI library, we can
remove the lib/icons.js file.

Change-Id: I735c01b41f970d304dc3aa06f1424863a09b0d20