Stop event propagation from search results click events
All Search Result components which can trigger the QuickView implement
the searchResult mixin. When links in these result components are
clicked, we want to ensure the Quickview is able to display regardless
of other event listeners that may be present (say on document.body).
These may come from a skin, or from a user script or gadget that's
running on the page.

For example, Commons currently runs this script
https://commons.wikimedia.org/wiki/MediaWiki:AnonymousI18N.js
for logged-out users who switch to a non-default language.

To ensure that search result links exhibit the intended behavior
(opening the quickview element in place rather than immediately
navigating to a new page), we need to stop the propogation of the
original event (generally a "click" event) that triggers the
showDetails() method. The custom "show-details" event that gets
fired by this method is sufficient to trigger the rest of the behavior
we want.

Bug: T260203
Change-Id: I8bde7b2285f10a4195e0182b4fddccffaad5a9af