Add select list component
- Adds a new Select component that replaces the native select
element. This is styled like the OOUI DropdownWidget and maintains
the same accessibility features.
- Changes the LookupResults component to SelectMenu so it can be
used for both the AutocompleteSearchInput and the Select component,
since they both use the same styles and functionality for their
dropdown menus
- Adds the ability for the SelectMenu to take in 3 different
item formats
- Adds a new named slot to the Tabs component for filters and adds
new styles for those filters, which are frameless Select components

Still to do: mobile styles for the filters. This will be handled
as part of the mobile UX task.

Bug: T258115
Change-Id: Id4b36929a5f784cda736f902953dae25780624a9