Replace infinite scroll in search results with "load more" button
For various reasons (inability to reach the page footer links,
performance issues, rate-limit errors, etc), loading search results
in a "true" infinite scroll fashion is not something we want to do.

This patch changes the behavior to autoload a set number of batches
of results (currently 2), and then display a button to "load more"
which the user must explicitly click to continue. Clicking the button
resets the counter and begins the cycle again (2 autoloads / 1 click).

This change requires adding a new piece of state to the App.vue
component, an "autoloadCounter" object which stores an integer count
value corresponding to each tab; auto-load count is tracked on a
per-tab basis. These values are set in the created() hook and are
reset whenever a new search is performed.

In addition to scrolling, switching tabs will also potentially load
more results, so it is possible for the counter for a given tab to
go below zero. Regardless of the actual value, once the counter is
zero or lower, infinite-scroll behavior stops and the load more
button is displayed.

Bug: T258516

Change-Id: I23c813c1fe0994e7dfb3932235f1f79b04601ef2