More rapidly decay item boost based on order
The current implementation decays the boost based on
the order of the match in a linear fashion, but that means
there's often little/negligeable difference between several
entities based on their order. The order is very important,
though, so let's decay much more rapidly so that the order
in which entities are returned matters much more.

Change-Id: I684d647b92c142291648dae0f0adb74290a18721