Improve QuickView accessibility
- Add a custom focus style for the close button, which gets focus
when the QuickView panel is opened
- Add visually-hidden text for the close button link
- Add the HTML5 hidden attribute to the QuickView aside element
when it's not open. We're still using 0 width to hide the element
for browsers that don't support hidden, but this will improve the
semantics for the aside in modern browsers
- Change the "More details" button to a link element styled like a
button. Though this is often considered an antipattern, this was
ultimately a design decision to give visual weight to this call
to action

Bug: T258110
Change-Id: I3e1bc45bb33d0ba866288acdfdcf233400afaec2