Clear the quickview of a given tab when that tab's filters change
Within a given tab, we want the quickview element to be immediately
cleared if the active filters for that tab change. Otherwise, the
result being previewed in the quickview might no longer exist in the
batch (a TIFF file when the user has filtered to see only JPEGs, etc).

By assigning a "ref" attribute (a key that matches the mediatype name)
to each SearchResults component, we allow the App component (which
handles filter-change events already) to programmatically call the
clearDetails() method on the appropriate SearchResults child whenever
a filter-change event is detected.

Quickview elements in other tabs will not be impacted.

Bug: T261818
Change-Id: I344540182aaace70015e303db1436dba75058b42