Align results to top
On a large enough screen with a small enough number of results
(= not filling the available height), opening quickview means
that the container for the thumbnails suddenly fills the entire
available height (because of quickview's 100vh), causing
thumbnails to spread out across the page. They need to stay up
top!

Bug: T261957
Change-Id: Id2be84720b631cf8ab8481acc0422a667ccc8ee3