Treat SelectMenu's formattedItems as a computed property
Bug T261883 describes a scenario where the user begins typing a search
term, sees the autocomplete suggestions populate, and then continues
typing (for example: first "port" and then completing to form
"portland").

In these situations the already-loaded autocomplete results were not
being replaced with new ones, even if the user deleted and re-typed a
completely different search term. New suggestions were only being
displayed if the user clicked out of the text input and then
re-focused back inside it.

This bug came from the fact that the SelectMenu component only generated
a set of "formattedItems" (which control what the user sees in this
dropdown) once, on first initialization. The results were stashed in
component state. Afterwards, the component was still receiving new
"items" props with new lookup results, but the "formattedItems" were
never updated. Correct results would only be displayed after the
component re-rendered due to focus changes.

The solution is to simply treat the formattedItems as a computed
property rather than state. This way changes to the "items" prop will
always result in a new set of "formattedItems".

One additional bug in the search input is also addressed by ensuring
that autocomplete suggestions are cleared if the user deletes all
input characters manually (instead of clicking the "x" button).

Bug: T261883
Change-Id: I5c57b9ea3a6144f720eee640b5d226b452f4ebcc