Add sort filter to all tabs
The sort filter has 3 potential values, all of which are handled
differently in the code:
1. Relevance is the default sort, so nothing is added to the search
query if this option is selected
2. Recency: the create_timestamp_desc sort option is added
3. Popularity: the popular_inclinks profile is used

This also adds a prefix prop to the base Select component that
adds text before the selected item (e.g. Sort By: for this use case).

Finally, since all tabs have a sort filter, the "hasFilters" computed
property has been removed from SearchFilters.vue since it is not
needed any longer.

Bug: T260722
Change-Id: I6d128fa4df6719967c3972792faed4305d1b57a9