Expand categories tab to include various namespaces
See first ticket for full list of namespaces. This patch updates the
"Categories" tab to "Categories and Pages" which includes results
from various non-file namespaces. Metadata is shown for categories
(but not for other namespaces as this data isn't included in
data returned by the search query).

This patch also updates the styles for text results (audio and
generic) to add a subtle border and to standardize layout and spacing.

Bug: T257700
Bug: T260392
Change-Id: I92c46213ccd56e05439e51fff7700d727d193000