Add a guard clause to resetCountAndLoadMore method
It's possible for the resetCountAndLoadMore() method of the App
component to be called in the absence of an actual search term (if the
user starts changing filter values before they have entered any text,
for example).

Adding a guard clause to this method that bails early if no search term
has been set will prevent invalid search requests from being made in
such cases.

Bug: T261536
Change-Id: I2e8b65bda2435b597cbfe51d16a56ef6fdc7f972