Add horizontal scrolling for filters on small screens
In order to show select list menus when overflow-x is set to auto,
we need to make the menu relative to a parent component, not to the
select list itself. Because of this, the menu will be 100% width.
This works well on mobile but becomes excessive on larger screens,
so this patch restricts horizontal scrolling to screen sizes below
the tablet breakpoint.

Bug: T258615
Change-Id: Ib3cb7e5942f795b26bbfddf1de266d85ecba629f