Increase statements boost based on amount of tokens
Full text scores will continue to grow in size as the
amount of tokens in the search term increases.
Statements OTOH are always just 1 token, even though they
represent the same multi-word concept.
As a result, the more tokens a search term contains, the
less relevant the statements became.

This will conditionally increase the statement boost
somewhat relative to the # of tokens in the search term.

Bug: T262522
Change-Id: I33f1fed1fec781756f8b701690ffb1aa250dd29c