Add concept chips UI
When a new search term is submitted (or when the App component is
created and there's a search term in the URL), fetch labels for
related concepts and display as concept chips. On click, the label
will become the new search term. Concept chips are only fetched
for the Images tab.

This feature requires the URL param conceptchips=true as a temporary
feature flag.

Bug: T256431
Change-Id: I394d31bed037ae6f1c7bed95b076a425f57d61eb