Set inheritAttrs to false where v-bind is used without arguments
For the SearchResult and QuickView components, we are passing in a
larger set of props than we typically need in any given use-case
because the exact data needed typically changes based on media type.
It's simpler to pass a single "result" object as a prop regardless of
media type and let the individual result components sort out what
information they need to display.

However, by default Vue.js will let un-used props passed to a component
"fall-through" as basic HTML attributes - even in cases where the name
of the property will collide with an existing HTML attribute like
"title" or "id".

Fortunately it is easy to change this behavior, just set
"inheritAttrs: false" in any component where this behavior is not
desired.

Bug: T261649
Change-Id: Iea1670c9d0650b2880781fb8ec4a41afe25cdad9